Refactor LWE lookup evaluators for split preimages
