+++ src/lookup/mod.rs
@@ -1,6 +1,7 @@
 pub mod commit_eval;
 mod ggh15;
 pub mod ggh15_eval;
+pub mod lwe;
 pub mod lwe_eval;
 pub mod poly;
 pub mod poly_vec;
